+++ tests/Aggregate/ReducerTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * Copyright MacFJA
  *
@@ -21,113 +23,125 @@ namespace Tests\MacFJA\RediSearch\Aggregate;
 
 use MacFJA\RediSearch\Aggregate\Reducer;
 use PHPUnit\Framework\TestCase;
+use Tests\MacFJA\RediSearch\support\Assertion;
 
 /**
- * @covers \MacFJA\RediSearch\Aggregate\Reducer::getQueryParts()
- * @covers \MacFJA\RediSearch\Aggregate\Reducer::__construct()
+ * @coversDefaultClass \MacFJA\RediSearch\Aggregate\Reducer
+ * @covers ::__construct
+ * @covers ::getQueryParts
+ *
  * @uses \MacFJA\RediSearch\Helper\DataHelper
+ *
+ * @see https://oss.redislabs.com/redisearch/Aggregations/#supported_groupby_reducers
  */
 class ReducerTest extends TestCase
 {
+    use Assertion;
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::count
+     * @covers ::count
      */
     public function testCount(): void
     {
-        self::assertReducerQuery('REDUCE COUNT 0', Reducer::count());
-        self::assertReducerQuery('REDUCE COUNT 0 AS countReducer', Reducer::count('countReducer'));
+        self::assertSameQuery('REDUCE COUNT 0', Reducer::count());
+        self::assertSameQuery('REDUCE COUNT 0 AS countReducer', Reducer::count('countReducer'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::countDistinct
+     * @covers ::countDistinct
      */
     public function testCountDistinct(): void
     {
-        self::assertReducerQuery('REDUCE COUNT_DISTINCT 1 @foobar', Reducer::countDistinct('foobar'));
-        self::assertReducerQuery('REDUCE COUNT_DISTINCT 1 @foobar AS countReducer', Reducer::countDistinct('foobar', 'countReducer'));
+        self::assertSameQuery('REDUCE COUNT_DISTINCT 1 @foobar', Reducer::countDistinct('foobar'));
+        self::assertSameQuery('REDUCE COUNT_DISTINCT 1 @foobar AS countReducer', Reducer::countDistinct('foobar', 'countReducer'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::countDistinctish
+     * @covers ::countDistinctish
      */
     public function testCountDistinctish(): void
     {
-        self::assertReducerQuery('REDUCE COUNT_DISTINCTISH 1 @foobar', Reducer::countDistinctish('foobar'));
-        self::assertReducerQuery('REDUCE COUNT_DISTINCTISH 1 @foobar AS countReducer', Reducer::countDistinctish('foobar', 'countReducer'));
+        self::assertSameQuery('REDUCE COUNT_DISTINCTISH 1 @foobar', Reducer::countDistinctish('foobar'));
+        self::assertSameQuery('REDUCE COUNT_DISTINCTISH 1 @foobar AS countReducer', Reducer::countDistinctish('foobar', 'countReducer'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::toList
+     * @covers ::toList
      */
     public function testToList(): void
     {
-        self::assertReducerQuery('REDUCE TOLIST 1 @foobar', Reducer::toList('foobar'));
-        self::assertReducerQuery('REDUCE TOLIST 1 @foobar AS list', Reducer::toList('foobar', 'list'));
+        self::assertSameQuery('REDUCE TOLIST 1 @foobar', Reducer::toList('foobar'));
+        self::assertSameQuery('REDUCE TOLIST 1 @foobar AS list', Reducer::toList('foobar', 'list'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::quantile
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::median
+     * @covers ::median
+     * @covers ::quantile
      */
     public function testQuantile(): void
     {
-        self::assertReducerQuery('REDUCE QUANTILE 2 @foobar 0.25', Reducer::quantile('foobar', 0.25));
-        self::assertReducerQuery('REDUCE QUANTILE 2 @foobar 0.5 AS median', Reducer::quantile('foobar', 0.5, 'median'));
-        self::assertReducerQuery('REDUCE QUANTILE 2 @foobar 0.5 AS median', Reducer::median('foobar', 'median'));
+        self::assertSameQuery('REDUCE QUANTILE 2 @foobar 0.25', Reducer::quantile('foobar', 0.25));
+        self::assertSameQuery('REDUCE QUANTILE 2 @foobar 0.5 AS median', Reducer::quantile('foobar', 0.5, 'median'));
+        self::assertSameQuery('REDUCE QUANTILE 2 @foobar 0.5 AS median', Reducer::median('foobar', 'median'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::average
+     * @covers ::average
      */
     public function testAverage(): void
     {
-        self::assertReducerQuery('REDUCE AVG 1 @foobar', Reducer::average('foobar'));
-        self::assertReducerQuery('REDUCE AVG 1 @foobar AS average', Reducer::average('foobar', 'average'));
+        self::assertSameQuery('REDUCE AVG 1 @foobar', Reducer::average('foobar'));
+        self::assertSameQuery('REDUCE AVG 1 @foobar AS average', Reducer::average('foobar', 'average'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::maximum
+     * @covers ::maximum
      */
     public function testMaximum(): void
     {
-        self::assertReducerQuery('REDUCE MAX 1 @foobar', Reducer::maximum('foobar'));
-        self::assertReducerQuery('REDUCE MAX 1 @foobar AS maximum', Reducer::maximum('foobar', 'maximum'));
+        self::assertSameQuery('REDUCE MAX 1 @foobar', Reducer::maximum('foobar'));
+        self::assertSameQuery('REDUCE MAX 1 @foobar AS maximum', Reducer::maximum('foobar', 'maximum'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::minimum
+     * @covers ::minimum
      */
     public function testMinimum(): void
     {
-        self::assertReducerQuery('REDUCE MIN 1 @foobar', Reducer::minimum('foobar'));
-        self::assertReducerQuery('REDUCE MIN 1 @foobar AS minimum', Reducer::minimum('foobar', 'minimum'));
+        self::assertSameQuery('REDUCE MIN 1 @foobar', Reducer::minimum('foobar'));
+        self::assertSameQuery('REDUCE MIN 1 @foobar AS minimum', Reducer::minimum('foobar', 'minimum'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::sum
+     * @covers ::sum
      */
     public function testSum(): void
     {
-        self::assertReducerQuery('REDUCE SUM 1 @foobar', Reducer::sum('foobar'));
-        self::assertReducerQuery('REDUCE SUM 1 @foobar AS sum', Reducer::sum('foobar', 'sum'));
+        self::assertSameQuery('REDUCE SUM 1 @foobar', Reducer::sum('foobar'));
+        self::assertSameQuery('REDUCE SUM 1 @foobar AS sum', Reducer::sum('foobar', 'sum'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::firstValue
+     * @covers ::firstValue
      */
     public function testFirstValue(): void
     {
-        self::assertReducerQuery('REDUCE FIRST_VALUE 1 @foobar', Reducer::firstValue('foobar'));
-        self::assertReducerQuery('REDUCE FIRST_VALUE 1 @foobar AS first', Reducer::firstValue('foobar', null, null, 'first'));
-        self::assertReducerQuery('REDUCE FIRST_VALUE 1 @foobar AS first', Reducer::firstValue('foobar', null, true, 'first'));
-        self::assertReducerQuery('REDUCE FIRST_VALUE 3 @foo BY @bar', Reducer::firstValue('foo', 'bar'));
-        self::assertReducerQuery('REDUCE FIRST_VALUE 4 @foo BY @bar ASC', Reducer::firstValue('foo', 'bar', true));
-        self::assertReducerQuery('REDUCE FIRST_VALUE 4 @foo BY @bar DESC', Reducer::firstValue('foo', 'bar', false));
-        self::assertReducerQuery('REDUCE FIRST_VALUE 4 @foo BY @bar DESC AS first', Reducer::firstValue('foo', 'bar', false, 'first'));
-        self::assertReducerQuery('REDUCE FIRST_VALUE 3 @foo BY @bar AS first', Reducer::firstValue('foo', 'bar', null, 'first'));
+        self::assertSameQuery('REDUCE FIRST_VALUE 1 @foobar', Reducer::firstValue('foobar'));
+        self::assertSameQuery('REDUCE FIRST_VALUE 1 @foobar AS first', Reducer::firstValue('foobar', null, null, 'first'));
+        self::assertSameQuery('REDUCE FIRST_VALUE 1 @foobar AS first', Reducer::firstValue('foobar', null, true, 'first'));
+        self::assertSameQuery('REDUCE FIRST_VALUE 3 @foo BY @bar', Reducer::firstValue('foo', 'bar'));
+        self::assertSameQuery('REDUCE FIRST_VALUE 4 @foo BY @bar ASC', Reducer::firstValue('foo', 'bar', true));
+        self::assertSameQuery('REDUCE FIRST_VALUE 4 @foo BY @bar DESC', Reducer::firstValue('foo', 'bar', false));
+        self::assertSameQuery('REDUCE FIRST_VALUE 4 @foo BY @bar DESC AS first', Reducer::firstValue('foo', 'bar', false, 'first'));
+        self::assertSameQuery('REDUCE FIRST_VALUE 3 @foo BY @bar AS first', Reducer::firstValue('foo', 'bar', null, 'first'));
     }
+
     /**
-     * @covers \MacFJA\RediSearch\Aggregate\Reducer::standardDeviation
+     * @covers ::standardDeviation
      */
     public function testStandardDeviation(): void
     {
-        self::assertReducerQuery('REDUCE STDDEV 1 @foobar', Reducer::standardDeviation('foobar'));
-        self::assertReducerQuery('REDUCE STDDEV 1 @foobar AS deviation', Reducer::standardDeviation('foobar', 'deviation'));
-    }
-
-    private static function assertReducerQuery(string $expected, Reducer $reducer): void
-    {
-        self::assertSame($expected, implode(' ', $reducer->getQueryParts()));
+        self::assertSameQuery('REDUCE STDDEV 1 @foobar', Reducer::standardDeviation('foobar'));
+        self::assertSameQuery('REDUCE STDDEV 1 @foobar AS deviation', Reducer::standardDeviation('foobar', 'deviation'));
     }
 }